restructure code, apply quick fixes, window pos saving
